requirements.txt
apache-airflow==2.6.0
numpy==1.24.1
pandas==1.5.2
scipy==1.10.0 
mlflow==2.2.1
scikit-learn==1.2.1
pyspark==3.3.2 
findspark==2.0.1
mlflow==2.2.1